Colocate tests with their code (#20)
Move all tests from centralized src/tests.rs to #[cfg(test)] modules
within their respective files. This improves code organization by:

- Making tests immediately visible when working on specific modules
- Better git history tracking (test changes with code changes)
- Following Rust best practices and project conventions
- Reducing merge conflicts

Tests migrated:
- src/iter.rs: text_nodes test
- src/parser.rs: 4 parser/serialization tests
- src/select.rs: 3 selector tests
- src/serializer.rs: 2 serializer tests

All 44 tests continue to pass.
